+++ inc/Core/Admin/Pages/Pipelines/assets/react/index.jsx
@@ -7,7 +7,6 @@
 import { render } from '@wordpress/element';
 import domReady from '@wordpress/dom-ready';
 import { QueryClientProvider } from '@tanstack/react-query';
-import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
 import { queryClient } from '@shared/lib/queryClient';
 import PipelinesApp from './PipelinesApp';
 import { HandlerProvider } from './context/HandlerProvider';
@@ -35,7 +34,6 @@ domReady( () => {
 			<HandlerProvider>
 				<PipelinesApp />
 			</HandlerProvider>
-			<ReactQueryDevtools initialIsOpen={false} />
 		</QueryClientProvider>,
 		rootElement
 	);